fix(db): reconcile instruments columns and cash_ledger_entries decimal type
Two fixes for the QNAP dev database:

1. KZO-83 modified migration 018 in-place after it was applied, leaving
   the instruments table missing type_raw, industry_category_raw, and
   finmind_date columns.

2. KZO-124 migration 020 converted trade_events/lots/lot_allocations
   price columns to NUMERIC(20,2) but missed cash_ledger_entries.amount
   (still INTEGER). Demo seed with ETF decimal prices (e.g. 185.50)
   produces fractional settlements that fail on INSERT with:
   "invalid input syntax for type integer: -576.5"

Also updates baseline schema for fresh database consistency.

+++ db/migrations/baseline_current_schema.sql
@@ -338,7 +338,7 @@ CREATE TABLE IF NOT EXISTS cash_ledger_entries (
       'REVERSAL'
     )
   ),
-  amount INTEGER NOT NULL CHECK (amount <> 0),
+  amount NUMERIC(20, 2) NOT NULL CHECK (amount <> 0),
   currency TEXT NOT NULL,
   related_trade_event_id TEXT REFERENCES trade_events(id) ON DELETE CASCADE,
   related_dividend_ledger_entry_id TEXT REFERENCES dividend_ledger_entries(id),
